requirements.txt
pandas
matplotlib.pyplot
seaborn
scipy